# docs/api/openapi.yaml
openapi: 3.0.0
info:
  title: Reelverse API
  version: 1.0.0
  description: API documentation for the Reelverse platform, including Account Abstraction, Paymaster, Upload, Billing, and Legal services.
servers:
  - url: /api/v1
    description: Main API server
...
paths:
  /api/v1/legal/dmca:
    post:
      summary: Submit DMCA Takedown Request
      tags:
        - Legal
      description: Allows users to submit a DMCA takedown request for infringing content.
      operationId: submitDmcaTakedown
      parameters:
        - $ref: '#/components/headers/X-Request-ID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DmcaTakedownRequest'
      responses:
        '201':
          description: DMCA takedown request submitted successfully.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/X-Request-ID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DmcaTakedownResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalServerError'
    get:
      summary: Get DMCA Takedown Request Details
      tags:
        - Legal
      description: Retrieves the details and current status of a specific DMCA takedown request.
      operationId: getDmcaTakedownDetails
      parameters:
        - name: id
          in: path
          required: true
          description: The unique ID of the DMCA takedown request.
          schema:
            type: string
            format: uuid
            example: "g1h2i3j4-k5l6-7890-1234-567890ghijkl"
        - $ref: '#/components/headers/X-Request-ID'
      responses:
        '200':
          description: DMCA takedown request details retrieved successfully.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/X-Request-ID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DmcaTakedownDetails'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          description: DMCA takedown request not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/InternalServerError'
...